+++ ChangeLog.md
@@ -3,6 +3,8 @@
 ## 0.8.32
 
 * Escape keys as necessary [#137](https://github.com/snoyberg/yaml/issues/137)
+* Support hexadecimal and octal number values [#135](https://github.com/snoyberg/yaml/issues/135)
+* More resilient `isNumeric` (should reduce cases of unneeded quoting)
 * hpackify
 * src subdir